fix: implement biblatex loading in citum-refs
Move biblatex entry conversion from citum-io into citum-refs so
the crate can parse .bib files directly. Replace the always-error
stub in formats::biblatex with a real implementation. citum-io
re-exports the module and delegates loading to citum-refs; remove
biblatex direct dep from citum-io.

Also add crates/citum-refs/README.md, fix unnecessary set_id.clone()
in load_merged_refs, and update crate docs to reflect BibLaTeX
is fully supported.

Resolves all five Copilot review comments on PR #815.
