+++ .beans/archive/csl26-p1iv--fix-pr-815-copilot-review-comments.md
@@ -1,0 +1,13 @@
+---
+# csl26-p1iv
+title: 'Fix PR #815 Copilot review comments'
+status: completed
+type: task
+priority: normal
+created_at: 2026-05-26T19:38:38Z
+updated_at: 2026-05-26T19:42:04Z
+---
+
+Move biblatex conversion logic from citum-io to citum-refs, implement the stub, fix clone, add README, update docs per Copilot review on PR #815.
+
+## Summary of Changes\n\nMoved biblatex conversion into citum-refs; replaced the stub with a real implementation; citum-io now re-exports and delegates. Removed biblatex dep from citum-io. Added README.md to citum-refs. Fixed unnecessary clone in load_merged_refs. Updated crate docs. All 25 tests pass.
