+++ crates/citum-io/Cargo.toml
@@ -12,7 +12,6 @@ citum-schema = { workspace = true, features = ["legacy-convert"] }
 citum-engine = { workspace = true }
 citum-refs = { workspace = true }
 csl-legacy = { workspace = true }
-biblatex = "0.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = { package = "serde_yaml_ng", version = "0.10" }
