+++ crates/citum-refs/README.md
@@ -1,0 +1,5 @@
+# citum-refs
+
+Multi-format bibliography reference loading and parsing for the Citum citation engine.
+
+Supports Citum YAML/JSON/CBOR, CSL-JSON, BibLaTeX, and RIS without depending on `citum-engine`.
