+++ crates/citum-refs/src/formats/biblatex.rs
@@ -5,6 +5,7 @@ SPDX-FileCopyrightText: © 2023-2026 Bruce D'Arcus and Citum contributors
 
 //! BibLaTeX bibliography loading helpers.
 
+use std::fs;
 use std::path::Path;
 
 use citum_schema::InputBibliography;
@@ -15,12 +16,17 @@ use crate::RefsError;
 ///
 /// # Errors
 ///
-/// Returns an error; BibLaTeX loading is not available in citum-refs due to circular
-/// dependency constraints. Use `citum-io::load_input_bibliography()` with
-/// `RefsFormat::Biblatex` instead.
-pub fn load_biblatex(_path: &Path) -> Result<InputBibliography, RefsError> {
-    Err(RefsError::ParseError(
-        "BibLaTeX".to_string(),
-        "BibLaTeX loading is not available in citum-refs. Use citum-io::load_input_bibliography() with RefsFormat::Biblatex instead.".to_string(),
-    ))
+/// Returns an error when the file cannot be read or parsed as BibLaTeX.
+pub fn load_biblatex(path: &Path) -> Result<InputBibliography, RefsError> {
+    let src = fs::read_to_string(path)?;
+    let bibliography = ::biblatex::Bibliography::parse(&src)
+        .map_err(|e| RefsError::ParseError("BibLaTeX".to_string(), e.to_string()))?;
+    let references = bibliography
+        .iter()
+        .map(crate::biblatex::input_reference_from_biblatex)
+        .collect();
+    Ok(InputBibliography {
+        references,
+        ..Default::default()
+    })
 }
